Update analysis file
Have fixed the bearing function, fixed the s.d. value that the prior/likelihood functions take and added more detail to the coordination. Have added a point to the plot to indicate the intersection point used to begin the simulation.

Next steps are to get the Metropolis-Hastings sampler loop actually working. The key is probably use loglikelihood and logprior functions, and therefore being able to change r so that the loop is not constantly being broken by zeros in the denominator. Also need to investigate whether the initialising values given in the material are appropriate as they seem to be causing issues and the bearing function which was given was broken.