Update analysis
Sample is now working correctly, although some parameters may still need to be adjusted. The snippet that was given to plot the sample has been added but is not yet working as expected. Next steps are to fix that, begin the write up and consider a case where there is alternate prior knowledge.